# docs/openapi.yaml
openapi: 3.1.0
info:
  title: zkAPI v2
  version: 2.0.0-testnet
  description: Native ETH lease, settlement and withdrawal Groth16 protocol API.
paths:
  /v2/billing/quote:
    get:
      summary: Read the pinned finalized native ETH billing quote
      responses:
        "200": {description: Native gwei quote with chain, feed and round binding}
  /v2/openrouter/leases:
    post:
      summary: Open a prompt-private, spending-limited OpenRouter lease
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApiRequestV2"}
      responses:
        "201":
          description: One-time plaintext runtime key and lease timing
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OpenRouterLease"}
        "400": {description: Invalid request, proof, or disabled lease mode}
        "409": {description: Existing active or settling lease}
        "429":
          description: OA key capacity is temporarily limited; the exact request is safe to retry
          headers:
            Retry-After:
              description: Seconds until the server recommends retrying the same request
              schema: {type: integer, minimum: 0}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /v2/openrouter/leases/{client_request_id}:
    get:
      summary: Read non-secret lease state; never returns the runtime key
      parameters:
        - in: path
          name: client_request_id
          required: true
          schema: {type: string}
      responses:
        "200": {description: Lease timing, status, and finalized aggregate usage}
        "404": {description: Unknown lease}
  /v2/withdraw/clearance:
    post:
      summary: Request a mutual-close clearance signature
      responses:
        "200": {description: Schnorr clearance signature}
  /v2/requests/{client_request_id}:
    get:
      summary: Recover a request transcript
      parameters:
        - in: path
          name: client_request_id
          required: true
          schema: {type: string}
      responses:
        "200": {description: Finalized transcript}
        "404": {description: Unknown request}
components:
  schemas:
    Groth16Proof:
      type: object
      required: [backend, proof]
      properties:
        backend: {const: groth16_bn254}
        proof: {type: string, contentEncoding: base64}
    ApiRequestV2:
      type: object
      required: [client_request_id, payload, payload_hash, public_inputs, proof]
      properties:
        client_request_id: {type: string}
        payload: {type: string}
        payload_hash: {type: string, pattern: "^0x[0-9a-f]+$"}
        public_inputs: {type: object, additionalProperties: true}
        proof: {$ref: "#/components/schemas/Groth16Proof"}
    ErrorResponse:
      type: object
      required: [status, client_request_id, error_code, error_message, retriable]
      properties:
        status: {const: error}
        client_request_id: {type: string}
        error_code:
          type: string
          description: Stable machine code; OA limits use oa_minute_request_limit, oa_hourly_issuance_budget, or oa_rate_limited.
        error_message: {type: string}
        retriable: {type: boolean}
        latest_root: {type: string, pattern: "^0x[0-9a-f]+$"}
        server_time_ms: {type: integer}
        retry_after_seconds:
          type: integer
          minimum: 0
          description: Present with OA rate limits and matches the Retry-After header.
    OpenRouterLease:
      type: object
      required: [status, client_request_id, api_key, openrouter_api_base, issued_at, expires_at, valid_for_seconds, settle_after, spending_limit_usd, key_source]
      properties:
        status: {const: active}
        client_request_id: {type: string}
        api_key: {type: string, description: Returned once and never persisted by zkapi-serverd}
        openrouter_api_base: {type: string, format: uri}
        issued_at: {type: integer}
        expires_at: {type: integer}
        valid_for_seconds: {type: integer}
        settle_after: {type: integer}
        spending_limit_usd: {type: number}
        key_source: {type: string, enum: [openrouter, oa_org]}
        verification:
          type: object
          description: Required when key_source is oa_org; submitted to the trusted OA verifier before inference.
          required: [verifier_url, station_id, station_recently_attested, key_valid_till, station_signature, org_signature]
          properties:
            verifier_url: {type: string, format: uri}
            station_id: {type: string}
            station_recently_attested: {type: boolean}
            key_valid_till: {type: integer}
            station_signature: {type: string}
            org_signature: {type: string}